+++ src/ripple/app/ledger/LedgerConsensus.h
@@ -17,8 +17,8 @@
 */
 //==============================================================================
 
-#ifndef RIPPLE_APP_CONSENSUS_LEDGERCONSENSUS_H_INCLUDED
-#define RIPPLE_APP_CONSENSUS_LEDGERCONSENSUS_H_INCLUDED
+#ifndef RIPPLE_APP_LEDGER_LEDGERCONSENSUS_H_INCLUDED
+#define RIPPLE_APP_LEDGER_LEDGERCONSENSUS_H_INCLUDED
 
 #include <ripple/app/ledger/Ledger.h>
 #include <ripple/app/ledger/LedgerProposal.h>
@@ -41,7 +41,7 @@ namespace ripple {
 class LedgerConsensus
 {
 public:
-    virtual ~LedgerConsensus() = 0;
+    virtual ~LedgerConsensus() = default;
 
     virtual Json::Value getJson (bool full) = 0;
 
@@ -66,17 +66,22 @@ public:
     virtual void simulate () = 0;
 };
 
-std::shared_ptr <LedgerConsensus>
-make_LedgerConsensus (
-    int previousProposers, int previousConvergeTime,
-    InboundTransactions& inboundTransactions, LocalTxs& localtx,
-    LedgerHash const & prevLCLHash, Ledger::ref previousLedger,
-        std::uint32_t closeTime, FeeVote& feeVote);
 
-void
-applyTransactions(std::shared_ptr<SHAMap> const& set, Ledger::ref applyLedger,
-                  Ledger::ref checkLedger,
-                  CanonicalTXSet& retriableTransactions, bool openLgr);
+//------------------------------------------------------------------------------
+/** Apply a set of transactions to a ledger
+
+  @param set                   The set of transactions to apply
+  @param applyLedger           The ledger to which the transactions should
+                               be applied.
+  @param checkLedger           A reference ledger for determining error
+                               messages (typically new last closed
+                                ledger).
+  @param retriableTransactions collect failed transactions in this set
+  @param openLgr               true if applyLedger is open, else false.
+*/
+void applyTransactions (std::shared_ptr<SHAMap> const& set,
+    Ledger::ref applyLedger, Ledger::ref checkLedger,
+    CanonicalTXSet& retriableTransactions, bool openLgr);
 
 } // ripple
