+++ src/ripple/app/misc/NetworkOPs.cpp
@@ -19,10 +19,11 @@
 
 #include <BeastConfig.h>
 #include <ripple/protocol/Quality.h>
-#include <ripple/app/ledger/LedgerConsensus.h>
 #include <ripple/core/DatabaseCon.h>
 #include <ripple/app/main/Application.h>
 #include <ripple/app/misc/FeeVote.h>
+#include <ripple/app/ledger/Consensus.h>
+#include <ripple/app/ledger/LedgerConsensus.h>
 #include <ripple/app/ledger/AcceptedLedger.h>
 #include <ripple/app/ledger/InboundLedger.h>
 #include <ripple/app/ledger/InboundLedgers.h>
@@ -123,17 +124,12 @@ public:
             deprecatedLogs().journal("FeeVote")))
         , mMode (omDISCONNECTED)
         , mNeedNetworkLedger (false)
-        , mProposing (false)
-        , mValidating (false)
         , m_amendmentBlocked (false)
         , m_heartbeatTimer (this)
         , m_clusterTimer (this)
+        , mConsensus (make_Consensus (*this))
         , m_ledgerMaster (ledgerMaster)
         , mCloseTimeOffset (0)
-        , lastCloseProposers_ (0)
-        , lastCloseConvergeTook_ (1000 * LEDGER_IDLE_INTERVAL)
-        , mLastCloseTime (0)
-        , mLastValidationTime (0)
         , mFetchPack ("FetchPack", 65536, 45, clock,
             deprecatedLogs().journal("TaggedCache"))
         , mFetchSeq (0)
@@ -157,8 +153,6 @@ private:
     std::uint32_t getCloseTimeNC (int& offset) const;
 
 public:
-    // Use *only* to timestamp our own validation.
-    std::uint32_t getValidationTimeNC () override;
     void closeTimeOffset (int) override;
 
     /** On return the offset param holds the System time offset in seconds.
@@ -219,15 +213,6 @@ public:
         std::uint32_t& minVal, std::uint32_t& maxVal) override
     {
         return m_ledgerMaster.getFullValidatedRange (minVal, maxVal);
-    }
-
-    STValidation::ref getLastValidation () override
-    {
-        return mLastValidation;
-    }
-    void setLastValidation (STValidation::ref v) override
-    {
-        mLastValidation = v;
     }
 
     //
@@ -329,8 +314,7 @@ public:
 
     bool recvValidation (
         STValidation::ref val, std::string const& source) override;
-    void takePosition (
-        int seq, std::shared_ptr<SHAMap> const& position) override;
+
     std::shared_ptr<SHAMap> getTXMap (uint256 const& hash);
     bool hasTXSet (
         const std::shared_ptr<Peer>& peer, uint256 const& set,
@@ -377,8 +361,6 @@ public:
     */
     void setStateTimer () override;
 
-    void newLCL (
-        int proposers, int convergeTime, uint256 const& ledgerHash) override;
     void needNetworkLedger () override
     {
         mNeedNetworkLedger = true;
@@ -395,44 +377,21 @@ public:
     {
         return !mNeedNetworkLedger && (mMode == omFULL);
     }
-    void setProposing (bool p, bool v) override
-    {
-        mProposing = p;
-        mValidating = v;
-    }
-    bool isProposing () override
-    {
-        return mProposing;
-    }
-    bool isValidating () override
-    {
-        return mValidating;
-    }
     bool isAmendmentBlocked () override
     {
         return m_amendmentBlocked;
     }
     void setAmendmentBlocked () override;
     void consensusViewChange () override;
-    std::uint32_t getLastCloseTime () override
-    {
-        return mLastCloseTime;
-    }
     void setLastCloseTime (std::uint32_t t) override
     {
-        mLastCloseTime = t;
+        mConsensus->setLastCloseTime (t);
     }
     Json::Value getConsensusInfo () override;
     Json::Value getServerInfo (bool human, bool admin) override;
     void clearLedgerFetch () override;
     Json::Value getLedgerFetchInfo () override;
     std::uint32_t acceptLedger () override;
-    Proposals & peekStoredProposals () override
-    {
-        return mStoredProposals;
-    }
-    void storeProposal (
-        LedgerProposal::ref proposal, RippleAddress const& peerPublic) override;
     uint256 getConsensusLCL () override;
     void reportFeeChange () override;
 
@@ -593,36 +552,18 @@ private:
     std::atomic<OperatingMode> mMode;
 
     std::atomic <bool> mNeedNetworkLedger;
-    bool mProposing;
-    bool mValidating;
     bool m_amendmentBlocked;
 
     beast::DeadlineTimer m_heartbeatTimer;
     beast::DeadlineTimer m_clusterTimer;
 
-    std::shared_ptr<LedgerConsensus> mConsensus;
-    NetworkOPs::Proposals mStoredProposals;
+    std::unique_ptr<Consensus> mConsensus;
+    std::shared_ptr<LedgerConsensus> mLedgerConsensus;
 
     LedgerMaster& m_ledgerMaster;
     InboundLedger::pointer mAcquiringLedger;
 
     int mCloseTimeOffset;
-
-    // The number of proposers who participated in the last ledger close
-    int lastCloseProposers_;
-
-    // How long the last ledger close took, in milliseconds
-    int lastCloseConvergeTook_;
-
-    // The hash of the last closed ledger
-    uint256 lastCloseHash_;
-
-    std::uint32_t mLastCloseTime;
-    std::uint32_t mLastValidationTime;
-    STValidation::pointer       mLastValidation;
-
-    // Recent positions taken
-    std::map<uint256, std::pair<int, std::shared_ptr<SHAMap>>> mRecentPositions;
 
     SubInfoMapType mSubAccount;
     SubInfoMapType mSubRTAccount;
@@ -741,11 +682,11 @@ void NetworkOPsImp::processHeartbeatTimer ()
         else if (mMode == omCONNECTED)
             setMode (omCONNECTED);
 
-        if (!mConsensus)
+        if (!mLedgerConsensus)
             tryStartConsensus ();
 
-        if (mConsensus)
-            mConsensus->timerEntry ();
+        if (mLedgerConsensus)
+            mLedgerConsensus->timerEntry ();
     }
 
     setHeartbeatTimer ();
@@ -805,10 +746,10 @@ std::string NetworkOPsImp::strOperatingMode () const
 
     if (mMode == omFULL)
     {
-        if (mProposing)
+        if (mConsensus->isProposing ())
             return "proposing";
 
-        if (mValidating)
+        if (mConsensus->isValidating ())
             return "validating";
     }
 
@@ -845,17 +786,6 @@ std::uint32_t NetworkOPsImp::getCloseTimeNC (int& offset) const
 {
     return iToSeconds (getNetworkTimePT (offset) +
                        boost::posix_time::seconds (mCloseTimeOffset));
-}
-
-std::uint32_t NetworkOPsImp::getValidationTimeNC ()
-{
-    std::uint32_t vt = getNetworkTimeNC ();
-
-    if (vt <= mLastValidationTime)
-        vt = mLastValidationTime + 1;
-
-    mLastValidationTime = vt;
-    return vt;
 }
 
 void NetworkOPsImp::closeTimeOffset (int offset)
@@ -1448,7 +1378,7 @@ void NetworkOPsImp::tryStartConsensus ()
         }
     }
 
-    if ((!mConsensus) && (mMode != omDISCONNECTED))
+    if ((!mLedgerConsensus) && (mMode != omDISCONNECTED))
         beginConsensus (networkClosed, m_ledgerMaster.getCurrentLedger ());
 }
 
@@ -1645,12 +1575,10 @@ bool NetworkOPsImp::beginConsensus (
             m_ledgerMaster.getClosedLedger ()->getHash ());
 
     // Create a consensus object to get consensus on this ledger
-    assert (!mConsensus);
+    assert (!mLedgerConsensus);
     prevLedger->setImmutable ();
 
-    mConsensus = make_LedgerConsensus (
-        lastCloseProposers_,
-        lastCloseConvergeTook_,
+    mLedgerConsensus = mConsensus->startRound (
         getApp().getInboundTransactions(),
         *m_localTX,
         networkClosed,
@@ -1664,7 +1592,7 @@ bool NetworkOPsImp::beginConsensus (
 
 bool NetworkOPsImp::haveConsensusObject ()
 {
-    if (mConsensus != nullptr)
+    if (mLedgerConsensus != nullptr)
         return true;
 
     if ((mMode == omFULL) || (mMode == omTRACKING))
@@ -1682,13 +1610,13 @@ bool NetworkOPsImp::haveConsensusObject ()
         {
             m_journal.info << "Beginning consensus due to peer action";
             if ( ((mMode == omTRACKING) || (mMode == omSYNCING)) &&
-                 (lastCloseProposers_ >= m_ledgerMaster.getMinValidations()) )
+                 (mConsensus->getLastCloseProposers() >= m_ledgerMaster.getMinValidations()) )
                 setMode (omFULL);
             beginConsensus (networkClosed, m_ledgerMaster.getCurrentLedger ());
         }
     }
 
-    return mConsensus != nullptr;
+    return mLedgerConsensus != nullptr;
 }
 
 uint256 NetworkOPsImp::getConsensusLCL ()
@@ -1696,7 +1624,7 @@ uint256 NetworkOPsImp::getConsensusLCL ()
     if (!haveConsensusObject ())
         return uint256 ();
 
-    return mConsensus->getLCL ();
+    return mLedgerConsensus->getLCL ();
 }
 
 void NetworkOPsImp::processTrustedProposal (
@@ -1717,42 +1645,20 @@ void NetworkOPsImp::processTrustedProposal (
         }
         else
         {
-            storeProposal (proposal, nodePublic);
+            mConsensus->storeProposal (proposal, nodePublic);
 
-            if (mConsensus->getLCL () == proposal->getPrevLedger ())
+            if (mLedgerConsensus->getLCL () == proposal->getPrevLedger ())
             {
-                relay = mConsensus->peerPosition (proposal);
+                relay = mLedgerConsensus->peerPosition (proposal);
                 m_journal.trace
                     << "Proposal processing finished, relay=" << relay;
             }
         }
 
         if (relay)
-            getApp().overlay().relay(*set,
-                proposal->getSuppressionID());
+            getApp().overlay().relay(*set, proposal->getSuppressionID());
         else
             m_journal.info << "Not relaying trusted proposal";
-    }
-}
-
-// Must be called while holding the master lock
-void
-NetworkOPsImp::takePosition (int seq, std::shared_ptr<SHAMap> const& position)
-{
-    mRecentPositions[position->getHash ()] = std::make_pair (seq, position);
-
-    if (mRecentPositions.size () > 4)
-    {
-        for (auto i = mRecentPositions.begin (); i != mRecentPositions.end ();)
-        {
-            if (i->second.first < (seq - 2))
-            {
-                mRecentPositions.erase (i);
-                return;
-            }
-
-            ++i;
-        }
     }
 }
 
@@ -1763,7 +1669,7 @@ NetworkOPsImp::mapComplete (uint256 const& hash,
     std::lock_guard<Application::MutexType> lock(getApp().getMasterMutex());
 
     if (haveConsensusObject ())
-        mConsensus->mapComplete (hash, map, true);
+        mLedgerConsensus->mapComplete (hash, map, true);
 }
 
 void NetworkOPsImp::endConsensus (bool correctLCL)
@@ -1782,7 +1688,7 @@ void NetworkOPsImp::endConsensus (bool correctLCL)
         }
     }
 
-    mConsensus = std::shared_ptr<LedgerConsensus> ();
+    mLedgerConsensus = std::shared_ptr<LedgerConsensus> ();
 }
 
 void NetworkOPsImp::consensusViewChange ()
@@ -2144,8 +2050,8 @@ bool NetworkOPsImp::recvValidation (
 
 Json::Value NetworkOPsImp::getConsensusInfo ()
 {
-    if (mConsensus)
-        return mConsensus->getJson (true);
+    if (mLedgerConsensus)
+        return mLedgerConsensus->getJson (true);
 
     Json::Value info = Json::objectValue;
     info[jss::consensus] = "none";
@@ -2204,23 +2110,23 @@ Json::Value NetworkOPsImp::getServerInfo (bool human, bool admin)
     info[jss::peers] = Json::UInt (getApp ().overlay ().size ());
 
     Json::Value lastClose = Json::objectValue;
-    lastClose[jss::proposers] = lastCloseProposers_;
+    lastClose[jss::proposers] = mConsensus->getLastCloseProposers();
 
     if (human)
     {
         lastClose[jss::converge_time_s] = static_cast<double> (
-            lastCloseConvergeTook_) / 1000.0;
+            mConsensus->getLastCloseDuration()) / 1000.0;
     }
     else
     {
         lastClose[jss::converge_time] =
-                Json::Int (lastCloseConvergeTook_);
+                Json::Int (mConsensus->getLastCloseDuration());
     }
 
     info[jss::last_close] = lastClose;
 
-    //  if (mConsensus)
-    //      info[jss::consensus] = mConsensus->getJson();
+    //  if (mLedgerConsensus)
+    //      info[jss::consensus] = mLedgerConsensus->getJson();
 
     if (admin)
         info[jss::load] = m_job_queue.getJson ();
@@ -2733,15 +2639,6 @@ bool NetworkOPsImp::unsubBook (std::uint64_t uSeq, Book const& book)
     return true;
 }
 
-void NetworkOPsImp::newLCL (
-    int proposers, int convergeTime, uint256 const& ledgerHash)
-{
-    assert (convergeTime);
-    lastCloseProposers_ = proposers;
-    lastCloseConvergeTook_ = convergeTime;
-    lastCloseHash_ = ledgerHash;
-}
-
 std::uint32_t NetworkOPsImp::acceptLedger ()
 {
     // This code-path is exclusively used when the server is in standalone
@@ -2756,19 +2653,8 @@ std::uint32_t NetworkOPsImp::acceptLedger ()
     beginConsensus (
         m_ledgerMaster.getClosedLedger ()->getHash (),
         m_ledgerMaster.getCurrentLedger ());
-    mConsensus->simulate ();
+    mLedgerConsensus->simulate ();
     return m_ledgerMaster.getCurrentLedger ()->getLedgerSeq ();
-}
-
-void NetworkOPsImp::storeProposal (
-    LedgerProposal::ref proposal, RippleAddress const& peerPublic)
-{
-    auto& props = mStoredProposals[peerPublic.getNodeID ()];
-
-    if (props.size () >= (unsigned) (lastCloseProposers_ + 10))
-        props.pop_front ();
-
-    props.push_back (proposal);
 }
 
 // <-- bool: true=added, false=already there
